moved all main stuff to its own components
